+++ src/test/java/org/scion/jpan/api/ScmpResponderTest.java
@@ -20,13 +20,14 @@ import java.io.IOException;
 import java.net.*;
 import java.nio.ByteBuffer;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.scion.jpan.*;
+import org.scion.jpan.testutil.ManagedThread;
+import org.scion.jpan.testutil.ManagedThreadNews;
 import org.scion.jpan.testutil.MockNetwork;
 import org.scion.jpan.testutil.MockScmpHandler;
 
@@ -57,18 +58,18 @@ public class ScmpResponderTest {
   }
 
   @Test
-  void testEcho() throws IOException, InterruptedException {
+  void testEcho() throws IOException {
     MockNetwork.startTiny();
     MockScmpHandler.stop(); // Shut down SCMP handler
     Path path = getPathTo112(InetAddress.getLoopbackAddress());
     // sender is in 110; responder is in 112
-    Thread responder = null;
+    ManagedThread responder = ManagedThread.newBuilder().build();
     try (ScmpSender sender = Scmp.newSenderBuilder().build()) {
       sender.setScmpErrorListener(scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
       sender.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
 
       // start responder
-      responder = startResponderThread(null);
+      responder.submit(mtn -> scmpResponder(null, mtn));
 
       // send request
       for (int i = 0; i < 10; i++) {
@@ -78,25 +79,25 @@ public class ScmpResponderTest {
         assertEquals(Scmp.TypeCode.TYPE_129, msg.getTypeCode());
       }
     } finally {
-      stopResponderThread(responder);
+      responder.stopNow();
       MockNetwork.stopTiny();
     }
   }
 
   @Test
-  void testEchoBlocked() throws IOException, InterruptedException {
+  void testEchoBlocked() throws IOException {
     MockNetwork.startTiny();
     MockScmpHandler.stop(); // Shut down SCMP handler
     Path path = getPathTo112(InetAddress.getLoopbackAddress());
     // sender is in 110; responder is in 112
-    Thread responder = null;
+    ManagedThread responder = ManagedThread.newBuilder().build();
     try (ScmpSender sender = Scmp.newSenderBuilder().build()) {
       sender.setScmpErrorListener(scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
       sender.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
 
       // start responder
       AtomicInteger dropCount = new AtomicInteger();
-      responder = startResponderThread((echoMsg) -> dropCount.incrementAndGet() < -42);
+      responder.submit(mtn -> scmpResponder(echoMsg -> dropCount.incrementAndGet() < -42, mtn));
 
       // send request
       sender.setTimeOut(100);
@@ -108,39 +109,20 @@ public class ScmpResponderTest {
       }
       assertEquals(2, dropCount.get());
     } finally {
-      stopResponderThread(responder);
+      responder.stopNow();
       MockNetwork.stopTiny();
     }
   }
 
-  private Thread startResponderThread(Predicate<Scmp.EchoMessage> predicate)
-      throws InterruptedException {
-    CountDownLatch barrier = new CountDownLatch(1);
-    Thread t = new Thread(() -> scmpResponder(barrier, predicate));
-    t.start();
-    barrier.await();
-    Thread.sleep(50);
-    return t;
-  }
-
-  private void stopResponderThread(Thread responder) throws InterruptedException {
-    if (responder != null) {
-      responder.join(100);
-      responder.interrupt(); // just in case.
-    }
-  }
-
-  private void scmpResponder(CountDownLatch barrier, Predicate<Scmp.EchoMessage> predicate) {
+  private void scmpResponder(Predicate<Scmp.EchoMessage> predicate, ManagedThreadNews mtn)
+      throws IOException {
     try (ScmpResponder responder = Scmp.newResponderBuilder().build()) {
       responder.setScmpErrorListener(
           scmpMessage -> errors.add(scmpMessage.getTypeCode().getText()));
       responder.setOption(ScionSocketOptions.SCION_API_THROW_PARSER_FAILURE, true);
       responder.setScmpEchoListener(predicate);
-      barrier.countDown();
+      mtn.reportStarted();
       responder.start();
-    } catch (IOException e) {
-      e.printStackTrace();
-      throw new RuntimeException(e);
     }
   }
 }